+++ jasperreports/src/net/sf/jasperreports/engine/util/JRBoxUtil.java
@@ -41,10 +41,10 @@ public final class JRBoxUtil
 	{
 		JRLineBox clone = box.clone(box.getBoxContainer());
 		
-		clone.setTopPadding(0);
-		clone.setLeftPadding(0);
-		clone.setBottomPadding(0);
-		clone.setRightPadding(0);
+		clone.setTopPadding((Integer)0);
+		clone.setLeftPadding((Integer)0);
+		clone.setBottomPadding((Integer)0);
+		clone.setRightPadding((Integer)0);
 		
 		//FIXMEBORDER does not copy padding correctly, if we only test line width
 		if (!keepLeft || box.getLeftPen().getLineWidth() <= 0f)
@@ -58,7 +58,7 @@ public final class JRBoxUtil
 			}
 			else
 			{
-				clone.getLeftPen().setLineWidth(0);
+				clone.getLeftPen().setLineWidth((Float)0f);
 			}
 		}
 		
@@ -73,7 +73,7 @@ public final class JRBoxUtil
 			}
 			else
 			{
-				clone.getRightPen().setLineWidth(0);
+				clone.getRightPen().setLineWidth((Float)0f);
 			}
 		}
 		
@@ -88,7 +88,7 @@ public final class JRBoxUtil
 			}
 			else
 			{
-				clone.getTopPen().setLineWidth(0);
+				clone.getTopPen().setLineWidth((Float)0f);
 			}
 		}
 		
@@ -103,7 +103,7 @@ public final class JRBoxUtil
 			}
 			else
 			{
-				clone.getBottomPen().setLineWidth(0);
+				clone.getBottomPen().setLineWidth((Float)0f);
 			}
 		}
 		
@@ -118,22 +118,22 @@ public final class JRBoxUtil
 	{
 		if (resetLeft)
 		{
-			box.getLeftPen().setLineWidth(0f);
+			box.getLeftPen().setLineWidth((Float)0f);
 		}
 		
 		if (resetRight)
 		{
-			box.getRightPen().setLineWidth(0f);
+			box.getRightPen().setLineWidth((Float)0f);
 		}
 
 		if (resetTop)
 		{
-			box.getTopPen().setLineWidth(0f);
+			box.getTopPen().setLineWidth((Float)0f);
 		}
 		
 		if (resetBottom)
 		{
-			box.getBottomPen().setLineWidth(0f);
+			box.getBottomPen().setLineWidth((Float)0f);
 		}
 	}
 	
@@ -233,14 +233,14 @@ public final class JRBoxUtil
 
 	public static void eraseBox(JRLineBox box)
 	{
-		box.setBottomPadding(0);
-		box.setTopPadding(0);
-		box.setLeftPadding(0);
-		box.setRightPadding(0);
-		box.getBottomPen().setLineWidth(0);
-		box.getTopPen().setLineWidth(0);
-		box.getLeftPen().setLineWidth(0);
-		box.getRightPen().setLineWidth(0);
+		box.setBottomPadding((Integer)0);
+		box.setTopPadding((Integer)0);
+		box.setLeftPadding((Integer)0);
+		box.setRightPadding((Integer)0);
+		box.getBottomPen().setLineWidth((Float)0f);
+		box.getTopPen().setLineWidth((Float)0f);
+		box.getLeftPen().setLineWidth((Float)0f);
+		box.getRightPen().setLineWidth((Float)0f);
 	}
 
 	private JRBoxUtil()
